CreateCodeCLI -> promptCodeCreator
