+++ packages/browser/src/context/QAWolfBrowserContext.ts
@@ -178,10 +178,6 @@ export class QAWolfBrowserContext extends EventEmitter {
 
   public async close() {
     if (CONFIG.create) {
-      logger.verbose("CREATE TIME!");
-      const {
-        CreateCodeCLI
-      } = require("/Users/jon/dev/examples/google/node_modules/qawolf/node_modules/@qawolf/cli/lib/CreateCodeCLI");
       await CreateCodeCLI.start({
         isTest: true,
         name: "google",
